Remove incorrect architectures field from library.properties
Previously, there were two architectures fields. The first correctly identified the library as being for all architectures. The second incorrectly identified the library as being for avr architecture only. The second field overrides the value of the first field. For this reason, I have removed the incorrect second architectures field.